pyproject.toml
[tool.poetry]
name = "mentosLab"
version = "0.1.0"
description = ""
authors = ["mukezhz <mukezhz@pm.me>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2.8"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.12.4"
psycopg2 = "^2.9.1"
djangorestframework-simplejwt = "^5.0.0"
python-dotenv = "^0.19.1"
django-cors-headers = "^3.10.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pynvim = "^0.4.3"
ipython = "^7.28.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
